+++ src/Games/RedScreen/RedScreen.cpp
@@ -6,7 +6,7 @@ void RedScreen::drawState() const
 {
     window_.clear();
     auto size = window_.getSize();
-    sf::RectangleShape rect(window_.mapPixelToCoords({size.x, size.y}));
+    sf::RectangleShape rect(window_.mapPixelToCoords({(int) size.x, (int)size.y}));
     rect.setFillColor(sf::Color::Red);
     window_.draw(rect);
     window_.display();
